refactor(scaffolding): extract BatchTestScaffolder collaborator
Move the batch (scaffold_test_batch_preview) and first-test-file
(scaffold_first_test_file_preview) orchestration out of the
ScaffoldingService facade into a dedicated internal BatchTestScaffolder
collaborator, mirroring the existing TypeScaffolder/SingleTestScaffolder
extractions. Content rendering stays in the shared static
TestScaffoldRenderer (consumed directly, never routed through
SingleTestScaffolder). The facade's logger-bound resolution helpers
(ResolveProject / ValidateIsTestProject / ResolveTestFramework) remain on
the facade — they are shared with the single-test path — and are supplied
to the collaborator as delegates rather than re-declared.

Behavior-preserving: the one-compilation-per-project batch cache and all
generated test-file output are unchanged, and IScaffoldingService's public
surface + DI lifetime are byte-identical. ScaffoldingIntegrationTests and
ScaffoldingFirstTestFileTests (33 tests) pass unmodified.

Closes: scaffolding-batch-first-test-collaborator-extraction
